Remove icon changing in breadcrumbs
